+++ plugins/org.eclipse.n4js.cli/src/org/eclipse/n4js/cli/N4jscFactory.java
@@ -13,7 +13,7 @@ package org.eclipse.n4js.cli;
 import org.eclipse.n4js.cli.compiler.N4jscLanguageClient;
 import org.eclipse.n4js.ide.N4JSIdeSetup;
 import org.eclipse.n4js.ide.xtext.server.XLanguageServerImpl;
-import org.eclipse.n4js.ide.xtext.server.XWorkspaceManager;
+import org.eclipse.n4js.ide.xtext.server.build.XWorkspaceManager;
 
 import com.google.common.base.Optional;
 import com.google.inject.Injector;
