+++ plugins/org.eclipse.n4js.ide/src/org/eclipse/n4js/ide/server/concurrent/N4JSQueuedExecutorService.java
@@ -13,7 +13,7 @@ package org.eclipse.n4js.ide.server.concurrent;
 import java.util.function.Function;
 
 import org.eclipse.n4js.ide.N4JSIdeDataCollectors;
-import org.eclipse.n4js.ide.xtext.server.concurrent.QueuedExecutorService;
+import org.eclipse.n4js.ide.xtext.server.QueuedExecutorService;
 import org.eclipse.n4js.smith.Measurement;
 import org.eclipse.xtext.util.CancelIndicator;
