+++ plugins/org.eclipse.n4js.ide/src/org/eclipse/n4js/ide/xtext/server/ResourceTaskContextAllContainerState.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.ide.xtext.server.openfiles;
+package org.eclipse.n4js.ide.xtext.server;
 
 import java.util.Collection;
 import java.util.Collections;
