+++ plugins/org.eclipse.n4js.ide/src/org/eclipse/n4js/ide/xtext/server/TextDocumentFrontend.java
@@ -63,15 +63,13 @@ import org.eclipse.lsp4j.jsonrpc.messages.Either;
 import org.eclipse.lsp4j.services.LanguageClient;
 import org.eclipse.lsp4j.services.LanguageClientExtensions;
 import org.eclipse.lsp4j.services.TextDocumentService;
-import org.eclipse.n4js.ide.xtext.server.concurrent.ConcurrentIndex;
-import org.eclipse.n4js.ide.xtext.server.concurrent.ConcurrentIndex.IIndexListener;
-import org.eclipse.n4js.ide.xtext.server.concurrent.ConcurrentIssueRegistry;
-import org.eclipse.n4js.ide.xtext.server.concurrent.ConcurrentIssueRegistry.IIssueRegistryListener;
-import org.eclipse.n4js.ide.xtext.server.concurrent.ConcurrentIssueRegistry.IssueRegistryChangeEvent;
+import org.eclipse.n4js.ide.xtext.server.ResourceTaskManager.IResourceTaskListener;
+import org.eclipse.n4js.ide.xtext.server.build.ConcurrentIndex;
+import org.eclipse.n4js.ide.xtext.server.build.ConcurrentIssueRegistry;
+import org.eclipse.n4js.ide.xtext.server.build.ConcurrentIndex.IIndexListener;
+import org.eclipse.n4js.ide.xtext.server.build.ConcurrentIssueRegistry.IIssueRegistryListener;
+import org.eclipse.n4js.ide.xtext.server.build.ConcurrentIssueRegistry.IssueRegistryChangeEvent;
 import org.eclipse.n4js.ide.xtext.server.contentassist.XContentAssistService;
-import org.eclipse.n4js.ide.xtext.server.openfiles.ResourceTaskContext;
-import org.eclipse.n4js.ide.xtext.server.openfiles.ResourceTaskManager;
-import org.eclipse.n4js.ide.xtext.server.openfiles.ResourceTaskManager.IResourceTaskListener;
 import org.eclipse.n4js.ide.xtext.server.rename.XIRenameService;
 import org.eclipse.n4js.xtext.workspace.IProjectConfigSnapshot;
 import org.eclipse.n4js.xtext.workspace.IWorkspaceConfigSnapshot;
