+++ tests/org.eclipse.n4js.ide.tests/src/org/eclipse/n4js/ide/tests/client/IdeTestLanguageClient.java
@@ -27,7 +27,7 @@ import org.eclipse.lsp4j.services.LanguageClient;
 import org.eclipse.n4js.ide.client.AbstractN4JSLanguageClient;
 import org.eclipse.n4js.ide.tests.server.AbstractIdeTest;
 import org.eclipse.n4js.ide.tests.server.StringLSP4J;
-import org.eclipse.n4js.ide.xtext.server.BuilderFrontend;
+import org.eclipse.n4js.ide.xtext.server.build.BuilderFrontend;
 import org.eclipse.n4js.projectModel.locations.FileURI;
 import org.eclipse.n4js.ts.scoping.builtin.N4Scheme;
 import org.eclipse.n4js.utils.URIUtils;
